Add gymnasium support to SAC (#378)
* Update sac_atari.py and sac_continuous_action.py to gymnasium's api

* Add testing

* https://github.com/vwxyzjn/cleanrl/pull/383/

* move test file

* fix final_info bug

* clean up mujoco tests

* update ci

* fix tests scripts

* Comment out test-mujoco-envs-mac

* fix final_observation

* test_pybullet.py

---------
