+++ tests/test_atari.py
@@ -15,11 +15,3 @@ def test_ppo_lstm():
         shell=True,
         check=True,
     )
-
-
-def test_sac():
-    subprocess.run(
-        "python cleanrl/sac_atari.py --learning-starts 10 --total-timesteps 16 --buffer-size 10 --batch-size 4",
-        shell=True,
-        check=True,
-    )
